Add support for decoding StringArray in LargeUtf8 schema
`Vec::<_>::from_type::<Record>(TracingOptions::default())?` is very
convenient, but it always returns a `LargeUtf8` type for String fields
in `Record`, which prevents using it before decoding from a
`RecordBatch` which may contain a `StringArray` (instead of the expected
`LargeStringArray`).

This change adds a fallback on error when decoding what we expect to be
a `LargeUtf8`, to try decoding it as a `Utf8` as a last resort.